refactor(cli): extract doctor gateway fallback
